add Discover and Notification page and cleaned VideoPage
This commit adds the discover page which queries for
the top trending Hashtags and presents them as scrollable
lists of videos. As such, the API has also been updated to
allow for fetching these hashtags. A placeholder notification page
has alos been added for later implementation.

The VideoPage has been reworked in order to fit the screen of the
device. This means removing constnt widths and heights, relying
on flexible layouts instead.

A new parameter labelled 'heroTag' must now be passed to a
VideoFeed in order to properly prefix the hero tag such that
only inter-related views trigger the hero animation.